refactor: move adapter classes to dedicated package and fix naming typos
- Move CompleteAdapter, PromptAdapter, and ResourceAdapter from org.springaicommunity.mcp.annotation to org.springaicommunity.mcp.adapter package
- Fix class name typos: PromptAdaptor → PromptAdapter, ResourceAdaptor → ResourceAdapter
- Update all import statements throughout the codebase to reflect the new package structure
- Reorganize imports for better code organization and consistency

This refactoring improves package organization by separating adapter utilities from annotation classes while maintaining the same public API.
